self_project/style_transfer/requirements.txt
Flask==1.1.1
gevent==1.4.0
h5py==2.10.0
numpy==1.17.0
Pillow==6.1.0
Werkzeug==0.16.0